fix narrow cast

+++ native/cocos/renderer/pipeline/custom/NativeRenderGraph.cpp
@@ -296,7 +296,7 @@ void addRasterViewImpl(
     auto &pass = get(RasterPassTag{}, passID, renderGraph);
     CC_EXPECTS(subpass.subpassID < num_vertices(pass.subpassGraph));
     auto &subpassData = get(SubpassGraph::SubpassTag{}, pass.subpassGraph, subpass.subpassID);
-    const auto slotID = subpass.rasterViews.size();
+    const auto slotID = static_cast<uint32_t>(subpass.rasterViews.size());
     CC_EXPECTS(subpass.rasterViews.size() == subpassData.rasterViews.size());
     auto nameIter = subpassData.rasterViews.find(name);
